Add `MovieDetailPosterView`
This view handles how to render a poster for the movie details screen.

Since the poster images have a vertical aspect ratio, but I don't want to use the full screen to show only
the poster, but other movie information as well, i'm using a "clever" cut to achieve the desired desing.

Now, since this view can be potentially used on iPad and the fact that text will be rendered on top that
depends on dynamic type, this view handles those cases as well