Avoid duplicating IDs when previewing movies.
Previously, the id was a random integer between `1` and `200`. Since the
top number was not too big, there was quite a chance that 2 of the 4
test elements would have the same ID, which resulted in a "missing" grid
element.

Then the solution is either use fixed ids on tests and/or make the
number bigger to lower the chance of 2 ids being the same for 2 test
values.