Add test plan
It's important to make the enviornment where the test are run as
predictably as possible.

While experimenting, I had issues on CI because the line height of text
was bigger on CI than on my local machine ([discussed on
github][discussion]).

The main factors I want to control in this test plan:

- Simulator language
- Simulator region

To match my local settings where the snapshot references where recorded

[discussion]: https://github.com/pointfreeco/swift-snapshot-testing/discussions/842